core training functions in separate module
